data-eng/requirements.txt
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
pyarrow>=12.0.0
